requirements.txt
pyyaml
Pillow
mastodon.py
pickledb
